requirements.txt
boto3==1.20.24
fastapi==0.89.1
mangum==0.12.3
uvicorn==0.17.1
Pillow==9.4.0
python-multipart==0.0.5
